refactor: add DeprecationNotice component and update import path in DetailHeader
