+++ web/app/components/plugins/base/deprecation-notice.tsx
